Remove artwiz-aleczapka from sbopkg-renames
I don't know what was the problem with this, but
I haven't found indication of a rename in the
slackbuilds.org mailing list or in the repository
itself.
For this reason, remove it from the sbopkg-renames
file.
Note that I still don't know the reason for the
old quirk, removed by r333.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>