Fix slack-desc parsing regression from r361
Prior to r361, there were two places where the slack-desc
was parsed: in the "package browser" and in the "search
results". r361 tossed most of the "package browser" code,
including that part.
It happens that the two implementation were different,
and of course the removed one was the better of the two.
This patch applies a further improved implementation to
the current code.

Thanks to "jsunx1" for providing the original
implementation and for pointing out the regression.
Thanks to Chess Griffin for preparing and sending me
a patch restoring the old behavior.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>