Update list: show progress indicator
Show a nice progress indicator instead of a "please wait"-style message.
The code containg a CLI implementation of a progress bar, and a gateway
function choosing between the "dialog" and the "cli" version depending
on whether we're using ncurses. The code should also be generic
enough to be reused whenever another gauge comes handy.

Many thanks to alkos333 for the idea and for the "dialog" implementation.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>