Don't always disable the GUI when CLI arguments are provided
Some CLI switches (-d, -f and -v) make sense even when
starting the GUI version of sbopkg.
Allow the usage of these flags also when starting the dialog
version of sbopkg.
Thanks to Patzy for raising the issue on #sbopkg and to
Chess Griffin for pointing out a weakness of the first version
of the patch.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>