Refactor the renames code and fix upgrading renamed packages.
This patch refactors the renames code putting the old/new
name retrieval in two generic helpers.
Also fix the upgrading of renamed packages by providing
the old package name to upgradepkg.
Thanks to Phillip Warner for raising the issue, and to
Chess Griffin and slakmagik for testing and reviewing this
patch.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>