Apply a huge cleanup patch.
This patch touches almost every part of sbopkg, making
it more readable, manageable and less bug-prone.
Working on this already let us spot several bugs, and
will ease further development and debugging efforts.

This is the result of a lot of patient team work, for
which Chess, slakmagik and I should be equally credited.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>