Implement renames tracking using a configuration directory.
Modify the renames tracking so that, instead of a single
file, all the files of a renames.d directory are used (just
like udev and modprobe do).
This has several advantages:
- no more need to "mv sbopkg-renames.new sbopkg-renames"
- it's now easier to add custom renames that automatically
  "survive" version upgrades
and, last but not least, this is believed to be the correct
way to handle these things.

Many thanks are due to Robby Workman for proposing this
change and to Chess Griffin for implementing it.