Use return values to implement the jump back to the main menu.
sbopkg 'touch'es a file (sbopkg_return_main_menu) to track
that the user asked to jump back to the main menu from a
"deep" menu level (such as info_item).
This patch converts it to the usage of function return values,
which makes the code smaller and easier to read.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>