get_source(): always check for md5 correctness.
David Spencer reported on the ML that the current sbopkg
behavior of checking the MD5 only just after the download
breaks some (bad) applications, like Google Earth, which
don't change the source file name on version upgrades.

This patch puts an additional check for existing sources,
so to workaround this issue (at the expense of extra MD5
checks for all other good applications).

See also r584 and r585 on -trunk.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>