Add option to retry a failed build.
This patch allows the user to retry a failed build. This
can be useful when the user mistakenly forgot to set up
some prerequisite (like creating a specific user or group),
when the build failed due to filling the build area (not
so unlikely when building large queues) and probably in
other cases too.
Addresses issue #24. Thanks to Zordrak for suggesting this
feature.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>