Properly validate repository descriptors.
Repository descriptors (*.repo files) weren't properly validated on
load. This could lead to unexpected problems whenever a wrong descriptor
is found.
This patch replaces the original 2-lines loader with a proper parsing
function which also performs some sanity checks.

While at it, stop using the magic number '7' as the number of fields in
a repository record -- use a variable instead.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>