Show currently installed package version in queue dialogs.
This patch shows the currently installed version of each queued package
in queue dialogs, or "New" if the package isn't installed.
The main reason to do something like this is to be quickly able to disable
the build of already-installed packages from a loaded queue. We're not
simply showing "Installed" here just to avoid inducing the user into
thinking that the queued package is exactly the same installed, while it
possibly isn't due to version updates.

Signed-off-by: Mauro Giachero <mauro.giachero@gmail.com>