fix failure to remove sources in get_source_names() ( issue 34 )
This is a band-aid which explicitly enables all flash-player-plugin sources to
be removed by testing for its names. It also fixes more general issues with
some SRCNAMEs not containing the VERSION number and get_source_names()
sometimes emitting bare newlines in cases where it should emit nothing at all.

This patch also modifies a regex in remove_sources_for_app() and adds unrelated
comments to string_search().

Thanks to artourter for the report, Chess for review, and Mauro for review and
suggestions.