Modified "Build options" dialog. Misc minor fixes.
Changed the menu order of the "Build options" dialog to default to using
"Saved" options rather than "None" on the theory that, if a user has
saved options, he will more likely want to use them than not.

Removed a commented-out block and made an assignment more consistent.
Brought the add_item_to_queue() comments more in line with its code.
Changed the filename value of GSNFILE so it gets caught by cleanup().
Fixed a typo.