Added a '?' argument to the -V flag and modified manpage and help output
This allows the user to invoke sbopkg with '-V ?' to get a list of valid
repo/branches and allows us to suggest that in the manpage and help output.
This was we don't have to keep up with manually itemizing the valid options.
Part of this change pops out a pre-existing block of code and turns it into
list_repos().