Renamed/revised remove_obsolete_packages() and added -O option
This revision converts remove_obsolete_packages() to
remove_uninstalled_packages() to better reflect what it currently
actually does and makes other minor revisions related to it. It also
adds a command line option (-O) to access this function. It's still
mnemonic with 'obsolete' because we're running out of option letters and
the function may yet be modified to remove obsolete sources.