add 60-SBo-current.repo; revise repo/renames files
Adding ponce's branch/fork thing of SBo for -current - it's supported as
far as including this to make it easier for people to use but no
further. Also got rid of most of the comments in the other files - there
are no user-serviceable parts inside and the filenames and descriptions
describe the files well enough, so this just keeps the key line. We'll
be linking the README in the repos.d directory so we can just say 'read
the README', especially as I recently changed the filename of the
README.