correct parameter expansion in get_source_names()
The assignment to SRCNAME was trying to cut out the longest match of
'%/download' rather than cutting off '/download'. The notable practical
effect was not recognizing that we had a source in the cache, since
foo.tar.gz/download != foo.tar.gz, so downloading it again. (Again, I
wonder how this could have ever gotten through testing and, again, I
feel like this sort of botched parameter expansion has happened before,
but can't find it in the vcs.)