add SWVER replaceable parameters
and additional VERSION parms. Now examples in sbopkg(8) and README-repos
which depend on Slackware/SBo versions won't have to choose between
being updated manually or looking out-of-date but will be updated
automatically. The sbopkg VERSIONs are just to be more consistent with
other usage - while svn users won't get to have manual pages saying
'SVN', the manuals' talking about an sbopkg/SVN dir were wrong and the
other uses were at least inconsistent and it's all vcs-specific anyway.
This revision is a tip of the iceberg to a larger backend release
process change. Which. will. happen. any. day... now....