restrict repo-based find commands to correct level
check_for_updates(), view_readmes(), search_package(), string_search(),
remove_obsolete_sources(): add mindepth/maxdepth flags (using a DEPTH
variable in string_search() (because it varies) and a FIND_FLAGS
variable in check_for_updates() (because it's reused)) to the find
commands that operate in the REPO. This will avoid matching things or
trying to use things like openvswitch/xen/foo when xen/foo is intended.
Thanks to David Somero for reporting this.