sbopkg: Move SRCNAME evaluation outside loop for performance
While building packages with a large number of sources (such as rust
packages), it was noticed that the source file MD5SUM verification was
slow.  This was due to setting and modifying the SRCNAME variable inside
the loop that checks the MD5SUM of every source file.  Because the
SRCNAME variable is loop invariant, move its evaluation to outside the
loop.  This drastically speeds up the verification phase.