sqg: Fix sqg detection of which repo it should use
If the master or current repos are used and someone switches back to
a stable SBo version, sqg will not use the stable version without
deleting the other repo(s) since it simply checks for the .git folder
first. If it is found, it never attempts to switch to the repo set
for sbopkg.

This change adds detection for the "master" and "current" repos and
will ensure sqg uses the proper repo that is set in /etc/sbopkg.conf
or /root/.sbopkg.conf.

Signed-off-by: Willy Sudiarto Raharjo <willysr@sbopkg.org>