Changed behaviour of use_options while running in bulk mode.
Added behaviour to use_options when running in bulk mode to
automatically prefer CLI-options over options specified in the
user's repository and prefer options specified in the user's
repository over options specified in a queuefile if one is used.

The man-page is also updated to document this behaviour.

Fixed indentation.

Signed-off-by: Willy Sudiarto Raharjo <willysr@sbopkg.org>